:wrench: Add interop between CommonJS and ESModule
